virtme_ng/utils: Remove cfg/virtme-ng.conf in favor of CONF_DEFAULT
The existing default configuration is problematic because new versions
might not correctly patch the existing user's files and depending on the
installation method adopted, it might not even be present. Therefore, move
the default into virtme_ng/utils.py by means of CONF_DEFAULT and use it as
a fallback to get_conf().

Signed-off-by: Ricardo B. Marlière <rbm@suse.com>